+++ src/structs.rs
@@ -125,6 +125,8 @@ impl Config {
                             t != &"description"
                                 && t != &"preimage"
                                 && t != &"sats_requested"
+                                && t != &"msats_requested"
+                                && t != &"msats_sent"
                                 && t != &"fee_msats"
                         })
                         .map(|s| s.to_string())
@@ -360,17 +362,15 @@ pub struct Pays {
     #[field_names_as_array(skip)]
     pub completed_at_str: String,
     pub payment_hash: String,
-    #[tabled(skip)]
-    #[field_names_as_array(skip)]
     pub msats_requested: u64,
     #[serde(skip_serializing)]
     pub sats_requested: u64,
-    #[tabled(skip)]
-    #[field_names_as_array(skip)]
     pub msats_sent: u64,
     #[serde(skip_serializing)]
     pub sats_sent: u64,
     pub fee_msats: u64,
+    #[serde(skip_serializing)]
+    pub fee_sats: u64,
     pub destination: String,
     #[serde(skip_serializing)]
     pub description: String,